FIX Ensure comments are escaped in RSS feeds.
FIX Also fix up preview to only output the comment content rather than the whole template.

FIX Hide preview after posting comment.

API Move AllowHtml to field to prevent issues with altering Html configuration after comments have been posted.

FIX If moderation is turned on for commenting, still render comments in preview mode.

+++ templates/CommentsInterface_singlecomment.ss
@@ -1,32 +1,30 @@
-<div class="comment" id="$Permalink">
-	<% if AllowHtml %>
-		$Comment.RAW
-	<% else %>
-		<p>$Comment.XML</p>
-	<% end_if %>
+<div class="comment" id="<% if isPreview %>comment-preview<% else %>$Permalink<% end_if %>">
+	$EscapedComment
 </div>
 
-<p class="info">
-	<% if $URL %>
-		<% _t('PBY','Posted by') %> <a href="$URL.URL" rel="nofollow">$AuthorName.XML</a>, $Created.Nice ($Created.Ago)
-	<% else %>
-		<% _t('PBY','Posted by') %> $AuthorName.XML, $Created.Nice ($Created.Ago)
-	<% end_if %>
-</p>
+<% if not isPreview %>
+	<p class="info">
+		<% if $URL %>
+			<% _t('PBY','Posted by') %> <a href="$URL.URL" rel="nofollow">$AuthorName.XML</a>, $Created.Nice ($Created.Ago)
+		<% else %>
+			<% _t('PBY','Posted by') %> $AuthorName.XML, $Created.Nice ($Created.Ago)
+		<% end_if %>
+	</p>
 
-<% if $ApproveLink || $SpamLink || $HamLink || $DeleteLink %>
-	<ul class="action-links">
-		<% if ApproveLink %>
-			<li><a href="$ApproveLink.ATT" class="approve"><% _t('APPROVE', 'approve this comment') %></a></li>
-		<% end_if %>
-		<% if SpamLink %>
-			<li><a href="$SpamLink.ATT" class="spam"><% _t('ISSPAM','this comment is spam') %></a></li>
-		<% end_if %>
-		<% if HamLink %>
-			<li><a href="$HamLink.ATT" class="ham"><% _t('ISNTSPAM','this comment is not spam') %></a></li>
-		<% end_if %>
-		<% if DeleteLink %>
-			<li class="last"><a href="$DeleteLink.ATT" class="delete"><% _t('REMCOM','remove this comment') %></a></li>
-		<% end_if %>
-	</ul>
+	<% if $ApproveLink || $SpamLink || $HamLink || $DeleteLink %>
+		<ul class="action-links">
+			<% if ApproveLink %>
+				<li><a href="$ApproveLink.ATT" class="approve"><% _t('APPROVE', 'approve this comment') %></a></li>
+			<% end_if %>
+			<% if SpamLink %>
+				<li><a href="$SpamLink.ATT" class="spam"><% _t('ISSPAM','this comment is spam') %></a></li>
+			<% end_if %>
+			<% if HamLink %>
+				<li><a href="$HamLink.ATT" class="ham"><% _t('ISNTSPAM','this comment is not spam') %></a></li>
+			<% end_if %>
+			<% if DeleteLink %>
+				<li class="last"><a href="$DeleteLink.ATT" class="delete"><% _t('REMCOM','remove this comment') %></a></li>
+			<% end_if %>
+		</ul>
+	<% end_if %>
 <% end_if %>
